display For Sale & Sold items on separate pages
When a user views the Sold section, clicks and views a Sold item, and then clicks the browser's back button, they may see the For Sale section instead of the Sold section, depending on the browser's caching. More logic would be required to fix this behavior, such as keeping track of the active tab, most likely with the URL's hash, and only loading the images in the active tab on initial page load. Having the items on separate pages instead of separate sections is a simpler solution.